feat: add set/toggle completion
